(MODULES-7463) Move all constants to V1Adapter
 - Only V1Adapter class is using the constants defined in the
   TaskScheduler2 class, so move them inline with the class

 - Remove any requires for taskscheduler2

 - Update constant references to use appropriate namespace

 - This allows taskscheduler2.rb to be deleted
