+++ lib/puppet_x/puppetlabs/scheduled_task/v1adapter.rb
@@ -1,6 +1,8 @@
 # This class is used to manage tasks using the Task Scheduler V2 API
 #
-require_relative './taskscheduler2'
+# https://msdn.microsoft.com/en-us/library/windows/desktop/aa383600(v=vs.85).aspx
+#
+require_relative './error'
 require_relative './trigger'
 
 module PuppetX
@@ -10,6 +12,95 @@ module ScheduledTask
 class V1Adapter
   # The name of the root folder for tasks
   ROOT_FOLDER = '\\'.freeze
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_enum_flags
+  class TASK_ENUM_FLAGS
+    TASK_ENUM_HIDDEN  = 0x1
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_action_type
+  class TASK_ACTION_TYPE
+    TASK_ACTION_EXEC          = 0
+    TASK_ACTION_COM_HANDLER   = 5
+    TASK_ACTION_SEND_EMAIL    = 6
+    TASK_ACTION_SHOW_MESSAGE  = 7
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_compatibility
+  # Win7/2008 R2                       = 3
+  # Win8/Server 2012 R2 or Server 2016 = 4
+  # Windows 10                         = 5 / 6
+  class TASK_COMPATIBILITY
+    TASK_COMPATIBILITY_AT     = 0
+    TASK_COMPATIBILITY_V1     = 1
+    TASK_COMPATIBILITY_V2     = 2
+    TASK_COMPATIBILITY_V2_1   = 3
+    TASK_COMPATIBILITY_V2_2   = 4
+    TASK_COMPATIBILITY_V2_3   = 5
+    TASK_COMPATIBILITY_V2_4   = 6
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_creation
+  class TASK_CREATION
+    TASK_VALIDATE_ONLY                 = 0x1
+    TASK_CREATE                        = 0x2
+    TASK_UPDATE                        = 0x4
+    # ( TASK_CREATE | TASK_UPDATE )
+    TASK_CREATE_OR_UPDATE              = 0x6
+    TASK_DISABLE                       = 0x8
+    TASK_DONT_ADD_PRINCIPAL_ACE        = 0x10
+    TASK_IGNORE_REGISTRATION_TRIGGERS  = 0x20
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_logon_type
+  class TASK_LOGON_TYPE
+    TASK_LOGON_NONE                           = 0
+    TASK_LOGON_PASSWORD                       = 1
+    TASK_LOGON_S4U                            = 2
+    TASK_LOGON_INTERACTIVE_TOKEN              = 3
+    TASK_LOGON_GROUP                          = 4
+    TASK_LOGON_SERVICE_ACCOUNT                = 5
+    TASK_LOGON_INTERACTIVE_TOKEN_OR_PASSWORD  = 6
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_run_flags
+  class TASK_RUN_FLAGS
+    TASK_RUN_NO_FLAGS             = 0
+    TASK_RUN_AS_SELF              = 0x1
+    TASK_RUN_IGNORE_CONSTRAINTS   = 0x2
+    TASK_RUN_USE_SESSION_ID       = 0x4
+    TASK_RUN_USER_SID             = 0x8
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_runlevel
+  class TASK_RUNLEVEL_TYPE
+    TASK_RUNLEVEL_LUA     = 0
+    TASK_RUNLEVEL_HIGHEST = 1
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_processtokensid
+  class TASK_PROCESSTOKENSID_TYPE
+    TASK_PROCESSTOKENSID_NONE           = 0
+    TASK_PROCESSTOKENSID_UNRESTRICTED   = 1
+    TASK_PROCESSTOKENSID_DEFAULT        = 2
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_state
+  class TASK_STATE
+    TASK_STATE_UNKNOWN    = 0
+    TASK_STATE_DISABLED   = 1
+    TASK_STATE_QUEUED     = 2
+    TASK_STATE_READY      = 3
+    TASK_STATE_RUNNING    = 4
+  end
+
+  # https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/ne-taskschd-_task_instances_policy
+  class TASK_INSTANCES_POLICY
+    TASK_INSTANCES_PARALLEL       = 0
+    TASK_INSTANCES_QUEUE          = 1
+    TASK_INSTANCES_IGNORE_NEW     = 2
+    TASK_INSTANCES_STOP_EXISTING  = 3
+  end
 
   public
   # Returns a new TaskScheduler object.
@@ -25,25 +116,25 @@ class V1Adapter
     @task_password = nil
 
     if compatibility_level == :v1_compatibility
-      self.compatibility = TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V1
+      self.compatibility = TASK_COMPATIBILITY::TASK_COMPATIBILITY_V1
     end
 
     set_account_information('',nil)
   end
 
   V1_COMPATIBILITY = [
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_AT,
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V1
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_AT,
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_V1
   ].freeze
 
   V2_COMPATIBILITY = [
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_4,
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_3,
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_2,
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_1,
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2,
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_AT,
-    TaskScheduler2::TASK_COMPATIBILITY::TASK_COMPATIBILITY_V1
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_4,
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_3,
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_2,
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2_1,
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_V2,
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_AT,
+    TASK_COMPATIBILITY::TASK_COMPATIBILITY_V1
   ].freeze
 
   # Returns an array of scheduled task names.
@@ -76,7 +167,7 @@ class V1Adapter
 
     task_folder = task_service.GetFolder(folder_path)
     filter_compatibility = !options[:include_compatibility].empty?
-    task_folder.GetTasks(TaskScheduler2::TASK_ENUM_FLAGS::TASK_ENUM_HIDDEN).each do |task|
+    task_folder.GetTasks(TASK_ENUM_FLAGS::TASK_ENUM_HIDDEN).each do |task|
       next if filter_compatibility && !options[:include_compatibility].include?(task.Definition.Settings.Compatibility)
       array << task.Path
     end
@@ -115,8 +206,8 @@ class V1Adapter
     task_password = nil
 
     case @definition.Principal.LogonType
-      when TaskScheduler2::TASK_LOGON_TYPE::TASK_LOGON_PASSWORD,
-        TaskScheduler2::TASK_LOGON_TYPE::TASK_LOGON_INTERACTIVE_TOKEN_OR_PASSWORD
+      when TASK_LOGON_TYPE::TASK_LOGON_PASSWORD,
+        TASK_LOGON_TYPE::TASK_LOGON_INTERACTIVE_TOKEN_OR_PASSWORD
         task_user = @definition.Principal.UserId
         task_password = @password
     end
@@ -124,7 +215,7 @@ class V1Adapter
     saved = task_folder.RegisterTaskDefinition(
       self.class.task_name_from_task_path(task_path),
       @definition,
-      TaskScheduler2::TASK_CREATION::TASK_CREATE_OR_UPDATE,
+      TASK_CREATION::TASK_CREATE_OR_UPDATE,
       task_user, task_password, @definition.Principal.LogonType)
 
     @task ||= saved
@@ -148,12 +239,12 @@ class V1Adapter
     if (user.nil? || user == "")
       # Setup for the local system account
       @definition.Principal.UserId = 'SYSTEM'
-      @definition.Principal.LogonType = TaskScheduler2::TASK_LOGON_TYPE::TASK_LOGON_SERVICE_ACCOUNT
-      @definition.Principal.RunLevel = TaskScheduler2::TASK_RUNLEVEL_TYPE::TASK_RUNLEVEL_HIGHEST
+      @definition.Principal.LogonType = TASK_LOGON_TYPE::TASK_LOGON_SERVICE_ACCOUNT
+      @definition.Principal.RunLevel = TASK_RUNLEVEL_TYPE::TASK_RUNLEVEL_HIGHEST
     else
       @definition.Principal.UserId = user
-      @definition.Principal.LogonType = TaskScheduler2::TASK_LOGON_TYPE::TASK_LOGON_PASSWORD
-      @definition.Principal.RunLevel = TaskScheduler2::TASK_RUNLEVEL_TYPE::TASK_RUNLEVEL_HIGHEST
+      @definition.Principal.LogonType = TASK_LOGON_TYPE::TASK_LOGON_PASSWORD
+      @definition.Principal.RunLevel = TASK_RUNLEVEL_TYPE::TASK_RUNLEVEL_HIGHEST
     end
 
     true
@@ -314,12 +405,12 @@ class V1Adapter
     action = nil
     (1..@definition.Actions.count).each do |i|
       index_action = action_at(i)
-      action = index_action if index_action.Type == TaskScheduler2::TASK_ACTION_TYPE::TASK_ACTION_EXEC
+      action = index_action if index_action.Type == TASK_ACTION_TYPE::TASK_ACTION_EXEC
       break if action
     end
 
     if action.nil? && create_if_missing
-      action = @definition.Actions.Create(TaskScheduler2::TASK_ACTION_TYPE::TASK_ACTION_EXEC)
+      action = @definition.Actions.Create(TASK_ACTION_TYPE::TASK_ACTION_EXEC)
     end
 
     action
